Tweaks: Changed the default search to Bing - seemingly better results.
Changed the default to use_thumbnails with an option to use  "use_raw_images" on the command line.
This only affect the Gallery, full quality images are downloaded in either case.
The raw images can take a minute or so to load, versus a few seconds when using thumbnails. While there is an improvement in image quality, it is not worth the wait.